Add inline videos, change urls and main nav, fixes
I added the videos inline and made them responsive. I changed the main
nav to have the longer names we talked about, and changed the URL
structure to match. I updated the homepage hero image to be the same
height as the videos. I added a workaround for a Chrome bug involving
rem units and font sizes.